==> Functions_Date_shift_models/Update_Dataframes/01_read_SEC_data.ipynb <==
import pandas as pd

import glob

from pathlib import Path
import csv

## Load pickle for exports and imports of data  
import pickle 
def load_obj(path):
    with open(path, 'rb') as f:
        return pickle.load(f)
    
def save_obj(obj, path ):
    with open(path, 'wb') as f:
        pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)

def read_ftd_file(file_path):
    df = pd.read_table(
    Path(file_path),
    sep="|",
    index_col=None
    )
    
    df = df.iloc[:-2 , :] ## Drop last two rows as they are useless to keep
    df.drop(['DESCRIPTION','PRICE'], axis=1,inplace=True)
    df.rename(columns={'SETTLEMENT DATE':'Date','QUANTITY (FAILS)':'QUANTITY_FAILS'},
             inplace=True)
    df['Date'] = pd.to_datetime(df['Date'])
    df.sort_values(["SYMBOL", "Date"], ascending = (True, True),inplace=True)
    df.set_index('Date',inplace=True)
    
    return df

## Do Loop to read all .txt files in folder 
## Requires import glob to do 
path = r'C:\Users\watso\LocalFolder\ProjectsLocal\Fintech_Project_2\SEC_Text' # USE YOUR PATH 
all_files = glob.glob(path + "/*.txt")

ftd_df = pd.DataFrame()   ## Declare empty dataframe 

loop_count = 0            ## Variables for catching errors in code - usually requires loading the SEC file in Excel, 
error_count = 0           ## locating the problem row, and fixing it in the SEC file itself 
error_list = []

for filename in all_files:
    if filename == 0:
        ftd_df = read_ftd_file(filename)
    else:
        ftd_df2 = read_ftd_file(filename)
        ftd_df = pd.concat([ftd_df,ftd_df2],axis='rows')
    loop_count+=1
    if ftd_df.isnull().values.any() == True:
        error_list.append(filename)
        error_count +=1

## Below code shows only 3 null values out of 7 million 
## Dropping the 3 null values

ftd_df.dropna(inplace=True)
ftd_df.sort_values(["SYMBOL", "Date"], ascending = (True, True),inplace=True)
ftd_df.reset_index(inplace=True)
ftd_df

filename

ftd_df.info()

ftd_df.isnull().values.any()

cusip_symbol_df = ftd_df[['SYMBOL','CUSIP']]
#cusip_symbol_df = cusip_symbol_df.drop_duplicates()
cusip_symbol_df = cusip_symbol_df.sort_values('SYMBOL')
cusip_symbol_df = cusip_symbol_df.reset_index(drop=True)
cusip_symbol_df = cusip_symbol_df.drop_duplicates()
cusip_symbol_df

cusip_symbol_df.isnull().values.any()

print(ftd_df[ftd_df.isnull().any(axis=1)]['QUANTITY_FAILS'].head())

print(ftd_df[ftd_df.isnull().any(axis=1)]['SYMBOL'].head())

## Export DFs
# ftd_df.to_csv('../Resources/ftd_all_data.csv')
# cusip_symbol_df.to_csv('../Resources/symbol_all_list.csv')
save_obj(ftd_df, '../Resources/ftd_all_data.pkl')
save_obj(cusip_symbol_df, '../Resources/symbol_all_list.pkl')
