==> Functions_Date_shift_models/Update_Dataframes/02_get_historical_data.ipynb <==
import pandas as pd

from pathlib import Path
import csv

import os
import requests
import json

import quandl

from dotenv import load_dotenv
load_dotenv()

## Constants 

## Set start date variable - dataframes will be created starting from this date
start_date = '2016-01-01'
end_date = '2021-12-31'
default_date_range = '71m' ## Default Range for IEX functions - don't need more at the moment

## QUANDL/NASDAQ 
nsdq_api_key = os.environ.get('NASDAQ_API_KEY')
base_url_nsdq = 'https://data.nasdaq.com/api/v3/datasets/FINRA/'

## Use pickle module to import and export and save files
import pickle
def load_obj(path):
    with open(path, 'rb') as f:
        return pickle.load(f)
def save_obj(obj, path ):
    with open(path, 'wb') as f:
        pickle.dump(obj, f, pickle.HIGHEST_PROTOCOL)

## Get Short "Interest" Data from Quandl 
def get_short_data_QUANDL(symbol):
    string_nsdq = "FINRA/FNSQ_"+symbol
    string_nyse = "FINRA/FNYX_"+symbol
    
    df1 = quandl.get(string_nsdq,start_date=start_date,end_date=end_date,authtoken=nsdq_api_key)   ## Nasdaq
    df2 = quandl.get(string_nyse,start_date=start_date,end_date=end_date,authtoken=nsdq_api_key)   ## NYSE

    df1 = df1.rename(columns={'ShortVolume':'ShortVolumeNSDQ','TotalVolume':'TotalVolumeNSDQ'})
    #df1 = df1.drop(columns={'ShortExemptVolume'})
    df1 = df1.rename(columns={'ShortExemptVolume':'ShortExemptVolumeNSDQ'})

    df2 = df2.rename(columns={'ShortVolume':'ShortVolumeNYSE','TotalVolume':'TotalVolumeNYSE'})
    #df2 = df2.drop(columns={'ShortExemptVolume'})
    df2 = df2.rename(columns={'ShortExemptVolume':'ShortExemptVolumeNYSE'})

    df3 = pd.merge(df1,df2,on='Date',how='outer')
    #df3 = df3.fillna(0)
    
    return df3


## Return FTD Data from SEC FTD files using a Stock's CUSIP number to sort 
def return_ftd_data_cusip(cusip_number):
    df = ftd_df.copy()
    df.set_index("CUSIP",inplace=True)
    df = df.loc[cusip_number]
    df['Date'] = pd.to_datetime(df['Date'])
    df = df.set_index('Date')
    return df

## Return the CUSIP symbol from the symbol_df symbol list 
def return_CUSIP_from_symbol(symbol):
    df = symbol_df.copy()
    df.set_index('SYMBOL',inplace=True)
    cusip_variable = df.loc[symbol]
    cusip_variable = cusip_variable['CUSIP']
    return cusip_variable

def return_ftd_data_symbol(symbol):
    cusip_number = return_CUSIP_from_symbol(symbol)
    df = return_ftd_data_cusip(cusip_number)
    return df

## FMP Constants 
fmpbase_urlv3 = 'https://fmpcloud.io/api/v3/'
fmpbase_urlv4 = 'https://fmpcloud.io/api/v4/'
api_key = os.getenv("FMP_CLOUD_API_KEY")

## FMP Functions 
def get_FMP_historical_data(symbol, startDate=start_date, endDate=end_date, apiKey=api_key):
    url_hist_price = fmpbase_urlv3+'historical-price-full/'
    url_hist_query_with_date = url_hist_price+symbol+'?from='+startDate+'&to='+endDate+'&apikey='+apiKey
    resp_data = requests.get(url_hist_query_with_date)
    json_ = resp_data.json()
    data = json_['historical']
    df = pd.DataFrame(data)
    df.rename(columns={'date':'Date'},inplace=True)
    df['Date'] = pd.to_datetime(df['Date'])
    df = df.reindex(index=df.index[::-1]) ## Reverse the DataFrame 
    df.set_index('Date',inplace=True)
    df.drop(columns='label',inplace=True)
    return df

def get_float_data_FMP(symbol):
    url_float_shares = fmpbase_urlv4+'shares_float?symbol='
    url_query_float_data = url_float_shares+symbol+'&apikey='+api_key
    resp_data = requests.get(url_query_float_data)
    #df = pd.DataFrame(resp_data.json())
    json_ = resp_data.json()
    return json_[0]

def get_company_profile_FMP_json(symbol):
    ## https://fmpcloud.io/api/v3/profile/AAPL?apikey='yourkeyhere'
    url_company_profile_url = fmpbase_urlv3+'profile/'+symbol+'?apikey='+api_key
    resp_data = requests.get(url_company_profile_url)
    json_response = resp_data.json()
    return json_response[0]

# def save_and_export_raw_df_csv(data, symbol, path='None'):
#     ## Can set custom path (useful for testing) otherwise will default to below path
#     if path=='None':
#         path = ('../FilesExportIndividualStockDFs_Big/'+symbol+'_combined_df.csv')
#     data.to_csv(path)
    
def save_and_export_raw_df_pkl(data, symbol, path='None'):
    ## Can set custom path (useful for testing) otherwise will default to below path
    if path=='None':
        path = ('../FilesExport_TimeSeries_DFs/'+symbol+'_combined_df.pkl')
    save_obj(data,path)

def get_time_series_data(symbol):
    ftd_data = return_ftd_data_symbol(symbol)
    ftd_data = ftd_data.drop(columns={'SYMBOL'})

    fmp_data = get_FMP_historical_data(symbol)
    df1 = pd.merge(fmp_data,ftd_data, on='Date',how='outer')
    df1['QUANTITY_FAILS'] = df1['QUANTITY_FAILS'].fillna(0)
    df1['volume'] = df1['volume'].fillna(0)
    df1['unadjustedVolume'] = df1['unadjustedVolume'].fillna(0)
    df1['vwap'] = df1['vwap'].fillna(0)
    df2 = get_short_data_QUANDL(symbol)
    df = pd.merge(df1,df2,on='Date',how='outer')
    df = df.fillna(0)
    return df

# ## Import FTD File using CSV 
# ftd_df = pd.read_csv(
#     Path('../Resources/ftd_all_data.csv'),
#     index_col=0, parse_dates=True
# )
# ## Import Symbol and CUSIP list using CSV
# symbol_df = pd.read_csv(
#     Path('../Resources/symbol_all_list.csv'),
#     index_col=0
# )


## Import dataframe objects using Pickle 
ftd_df = load_obj('../Resources/ftd_all_data.pkl')
symbol_df = load_obj('../Resources/symbol_all_list.pkl')

## Rewrite loop to access FMP data using list of symbols from symbol_df 
## This code is written this way, with the while loop, because it allows you to pick up from close to
## where you left off, in the event of an API failure, or the code stopping. 
## Changing the x and y starting values will allow you to pick up from where the loop failed, and continue
## reading data. 

## Create while loop that access FMP and gets data 
x = 0    
y = 50
increment = 50  ## Make sure increment is right

## Setup in 50 increments to find error in case the loop fails, but keep previous data.

# ## Initialize lists for successful calls
# symbol_success_list = []
# symbol_not_success_list = [] 

length_ = len(symbol_df)
test_length = 200 
loop_length = length_ ## Set to length_ to run as full, or your test_length

## Reimport lists 
symbol_success_list = load_obj('../Resources/02_symbol_success_list.pkl')
len(symbol_success_list)

symbol_not_success_list = load_obj('../Resources/02_symbol_not_success_list.pkl')
len(symbol_not_success_list)

#symbol_export_list[3497]

## Code to locate index of symbol 
test_symbol = 'LMDXW'

length_ = len(symbol_df)
index_variable = 0 

for i in range(length_):
    if symbol_df['SYMBOL'].iloc[i] == test_symbol:
        index_variable = i
        print(symbol_df['SYMBOL'].iloc[i])
        print(index_variable)
        break

x = 18794    
y = x + 50

while (x < loop_length):

    ## Set y value for end of while loop
    if y >= length_: y = length_
    ## In theory this should stop the for loop from crashing 
    ##    at the end and running out of index range     
    
    range_var = range(x,y)
    ## Set label variables for exports 
    str_symbol1 = str(x)
    y2 = y - 1 
    str_symbol2 = str(y2)
    
    for i in range_var: 
        
        ## Create temp lists 
        temp_symbol_success_list = []
        temp_symbol_not_success_list = []
        
        ## Iterate through symbol list and create data by symbol
        symbol_var = symbol_df['SYMBOL'][i]
        
        # Use while running loop, if error is made, can see what index it happened on 
        # can also check current value of symbol_var as well 
        error_var = i
        
        ## Get Time Series Data 
        try:
            time_series_df = get_time_series_data(symbol_var)
            ## Export as a list with a key (the stock symbol) to the dataframe 
            export_obj = [{symbol_var:time_series_df}]
            
            if len(time_series_df) > 1500:
                symbol_success_list.append(symbol_var)
            else:
                symbol_not_success_list.append(symbol_var)               
            
        except:
            continue
            
            
        ## If successful, export data 
        save_and_export_raw_df_pkl(export_obj,symbol_var)
            
    ## If for loop successful, append temp_symbol lits to main lists 
    symbol_success_list.extend(temp_symbol_success_list)
    symbol_not_success_list.extend(temp_symbol_not_success_list)
    
    ## Export main lists each time, and rewrite during each iteration of while loop 
    ## Can read in list afterwards, to figure out where loop went wrong, and where to restart 
    save_obj(symbol_success_list,'../Resources/02_symbol_success_list.pkl')
    save_obj(symbol_not_success_list,'../Resources/02_symbol_not_success_list.pkl')
    
    ## Check before run, if incorrect, can waste a lot of API credits  
    x += increment
    y += increment
